backend/requirements.txt
asgiref==3.8.1
attrs==24.1.0
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-ranged-response==0.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.4.0
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
